feat: centralize settings management
